Move interface/host management into Worker class
