+++ distributed/utils_test.py
@@ -501,7 +501,7 @@ def run_worker(q, scheduler_q, **kwargs):
         with pristine_loop() as loop:
             scheduler_addr = scheduler_q.get()
             worker = Worker(scheduler_addr, validate=True, **kwargs)
-            loop.run_sync(lambda: worker._start(0))
+            loop.run_sync(lambda: worker._start())
             q.put(worker.address)
             try:
 
@@ -521,7 +521,7 @@ def run_nanny(q, scheduler_q, **kwargs):
         with pristine_loop() as loop:
             scheduler_addr = scheduler_q.get()
             worker = Nanny(scheduler_addr, validate=True, **kwargs)
-            loop.run_sync(lambda: worker._start(0))
+            loop.run_sync(lambda: worker._start())
             q.put(worker.address)
             try:
                 loop.start()
