+++ distributed/worker.py
@@ -29,7 +29,7 @@ from tornado.locks import Event
 from . import profile, comm
 from .batched import BatchedSend
 from .comm import get_address_host, get_local_address_for, connect
-from .comm.utils import offload
+from .comm.utils import offload, uri_from_host_port
 from .compatibility import unicode, get_thread_identity, finalize, MutableMapping
 from .core import error_message, CommClosedError, send_recv, pingpong, coerce_to_address
 from .diskutils import WorkSpace
@@ -45,6 +45,7 @@ from .threadpoolexecutor import ThreadPoolExecutor, secede as tpe_secede
 from .utils import (
     funcname,
     get_ip,
+    get_ip_interface,
     has_arg,
     _maybe_complex,
     log_errors,
@@ -294,6 +295,9 @@ class Worker(ServerNode):
         extensions=None,
         metrics=None,
         data=None,
+        interface=None,
+        host=None,
+        port=None,
         low_level_profiler=dask.config.get("distributed.worker.profile.low-level"),
         **kwargs
     ):
@@ -405,7 +409,20 @@ class Worker(ServerNode):
             scheduler_addr = coerce_to_address(scheduler_ip)
         else:
             scheduler_addr = coerce_to_address((scheduler_ip, scheduler_port))
-        self._port = 0
+        self.contact_address = contact_address
+
+        if interface:
+            if host:
+                raise ValueError("Can not specify both interface and host")
+            else:
+                host = get_ip_interface(interface)
+
+        if host or port:
+            self._start_addr = uri_from_host_port(host, port, 0)
+        else:
+            # Choose appropriate address for scheduler
+            self._start_addr = None
+
         self.ncores = ncores or _ncores
         self.total_resources = resources or {}
         self.available_resources = (resources or {}).copy()
@@ -416,7 +433,6 @@ class Worker(ServerNode):
         self.preload_argv = preload_argv
         if self.preload_argv is None:
             self.preload_argv = dask.config.get("distributed.worker.preload-argv")
-        self.contact_address = contact_address
         self.memory_monitor_interval = parse_timedelta(
             memory_monitor_interval, default="ms"
         )
@@ -885,6 +901,7 @@ class Worker(ServerNode):
     @gen.coroutine
     def _start(self, addr_or_port=0):
         assert self.status is None
+        addr_or_port = addr_or_port or self._start_addr
 
         enable_gc_diagnosis()
         thread_state.on_event_loop_thread = True
